refactor(ui): split chat page responsibilities
